telegram_chupakabrada_bot/requirements.txt
beautifulsoup4
black
bs4
isort
markovify
pylint
pyTelegramBotAPI
python-dotenv
requests
tabulate
deptry
